requirements.txt
# keep this list tiny on purpose
streamlit>=1.30.0
python-docx>=1.1.0
